18th commit (Refactor: Moving extra logic from main.py to separate files)
