+++ app/templates/index.html
@@ -1,0 +1,8 @@
+{% extends "base.html" %}
+
+{% block title %}Home{% endblock %}
+
+{% block content %}
+    <h2>Welcome to the Home Page</h2>
+    <p>This is the home page of the FastAPI E-commerce application.</p>
+{% endblock %}
