+++ app/templates/logout.html
@@ -1,0 +1,8 @@
+{% extends "base.html" %}
+
+{% block title %}Logout{% endblock %}
+
+{% block content %}
+    <h2>You have been logged out</h2>
+    <p>Thank you for visiting FastAPI E-commerce. <a href="/">Return to Home</a>.</p>
+{% endblock %}
