+++ app/templates/profile.html
@@ -1,0 +1,9 @@
+{% extends "base.html" %}
+
+{% block title %}Profile{% endblock %}
+
+{% block content %}
+    <h2>Welcome, {{ user_id }}</h2>
+    <p>This is your profile page.</p>
+    <a href="/logout">Logout</a>
+{% endblock %}
